+++ src/async/export/TokenBucket.js
@@ -1,8 +1,8 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+import { IntfTimeSource, StdTimeSource } from '@this/clocks';
 import { Duration, Frequency, Moment } from '@this/data-values';
-import { IntfTimeSource, StdTimeSource } from '@this/metacomp';
 import { MustBe } from '@this/typey';
 
 import { ManualPromise } from '#x/ManualPromise';