+++ src/loggy/export/StdLoggingEnvironment.js
@@ -1,10 +1,9 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
-import * as process from 'node:process';
-
 import { EventSource } from '@this/async';
-import { Moment, StackTrace } from '@this/data-values';
+import { WallClock } from '@this/clocks';
+import { StackTrace } from '@this/data-values';
 import { MustBe } from '@this/typey';
 
 import { BaseLoggingEnvironment } from '#x/BaseLoggingEnvironment';
@@ -20,15 +19,6 @@ export class StdLoggingEnvironment extends BaseLoggingEnvironment {
 
   /** @type {IdGenerator} ID generator to use. */
   #idGenerator = new IdGenerator();
-
-  /** @type {bigint} Last result from `hrtime.bigint()`. */
-  #lastHrtimeNsec = -1n;
-
-  /**
-   * @type {bigint} Last "now" measured by {@link #now}, as a `bigint`
-   * representing a nanosecond-based Unix Epoch time.
-   */
-  #lastNowNsec = -1n;
 
   /**
    * Constructs an instance.
@@ -48,7 +38,7 @@ export class StdLoggingEnvironment extends BaseLoggingEnvironment {
 
   /** @override */
   _impl_makeId() {
-    return this.#idGenerator.makeId(this.#now());
+    return this.#idGenerator.makeId(WallClock.now());
   }
 
   /** @override */
@@ -59,60 +49,6 @@ export class StdLoggingEnvironment extends BaseLoggingEnvironment {
 
   /** @override */
   _impl_now() {
-    return this.#now();
+    return WallClock.now();
   }
-
-  /**
-   * Gets the "now" moment as a plain number of seconds since the Unix Epoch.
-   *
-   * @returns {Moment} "Now."
-   */
-  #now() {
-    // What's going on here: We attempt to use `hrtime()` -- which has nsec
-    // precision but an arbitrary zero-time, and which we don't assume runs at
-    // exactly (effective) wall-clock rate -- to improve on the precision of
-    // `Date.now()` -- which has msec precision and a well-established base, and
-    // which we assume is as accurate as it is precise.
-
-    const hrtimeNsec  = process.hrtime.bigint();
-    const dateNowNsec = BigInt(Date.now()) * StdLoggingEnvironment.#MSEC_PER_NSEC;
-    let nowNsec;
-
-    if (this.#lastNowNsec < 0) {
-      nowNsec = dateNowNsec;
-    } else {
-      const hrDiffNsec    = hrtimeNsec - this.#lastHrtimeNsec;
-      const hrTrackedNsec = this.#lastNowNsec + hrDiffNsec;
-      if (   (hrTrackedNsec >= (dateNowNsec - StdLoggingEnvironment.#MSEC_PER_NSEC))
-          && (hrTrackedNsec <= (dateNowNsec + StdLoggingEnvironment.#MSEC_PER_NSEC))) {
-        nowNsec = hrTrackedNsec;
-      } else {
-        // The wall time reconstructed from the difference between `hrtime()`
-        // readings is too far off from what `Date.now()` reports. That is, it's
-        // not useful, so we just use a straight `Date.now()` result.
-        nowNsec = dateNowNsec;
-      }
-    }
-
-    if (nowNsec < this.#lastNowNsec) {
-      nowNsec = this.#lastNowNsec + StdLoggingEnvironment.#MSEC_PER_NSEC;
-    }
-
-    this.#lastHrtimeNsec  = hrtimeNsec;
-    this.#lastNowNsec     = nowNsec;
-
-    const nowSec = Number(nowNsec) * StdLoggingEnvironment.#SECS_PER_NSEC;
-    return new Moment(nowSec);
-  }
-
-
-  //
-  // Static members
-  //
-
-  /** @type {number} The number of seconds in a nanosecond. */
-  static #SECS_PER_NSEC = 1 / 1_000_000_000;
-
-  /** @type {bigint} The number of milliseconds in a nanosecond. */
-  static #MSEC_PER_NSEC = 1_000_000n;
 }