+++ CHANGELOG.md
@@ -7,9 +7,15 @@ versioning principles. Unstable releases do not.
 ### [Unreleased]
 
 Breaking changes:
-* None.
+* `clocks`:
+  * New module!
+  * Moved time-related classes to here from `metacomp`.
+  * New class `WallClock`, extracted from `StdLoggingEnvironment`, so that other
+    stuff can get sub-msec wall times.
 
 Other notable changes:
+* `builtin-services`:
+  * `RequestLogger` now logs sub-msec request durations.
 * `net-util`:
   * Added a bit to `OutgoingResponse` to add detail when it would otherwise try
     to double-resolve its "when-completed" promise. This seems to be happening